+++ src/routes/upload.ts
@@ -1,10 +1,10 @@
 import { Router } from 'express'
 import { storage } from '../utils/multer'
-import { uploadFileControllerHandler } from '../controllers/add-file'
-import { uploadFilesControllerHandler } from '../controllers/add-multiples-files'
-import { deleteFileControllerHandler } from '../controllers/delete-file'
-import { listFilesControllerHandler } from '../controllers/list-all-files'
-import { listFileByIdControllerHandler } from '../controllers/list-file-by-id'
+import { uploadFileControllerHandler } from '../controllers/upload-file-controller'
+import { uploadFilesControllerHandler } from '../controllers/multi-upload-controller'
+import { deleteFileControllerHandler } from '../controllers/delete-file-controller'
+import { listFilesControllerHandler } from '../controllers/list-all-files-controller'
+import { listFileByIdControllerHandler } from '../controllers/list-file-by-id-controller'
 import { fileValidatorMiddleware } from '../middleware/validate-files'
 import { Request, Response } from 'express'
 export const upload = Router()
